+++ R/read_delim.R
@@ -9,7 +9,11 @@ NULL
 #' flat file data, comma separated values and tab separated values,
 #' respectively. \code{read_csv2} uses \code{;} for separators, instead of
 #' \code{,}. This is common in European countries which use \code{,} as the
-#' decimal separator.
+#' decimal separator. The \code{spec_*} versions return the column
+#' specification instead of the full data frame.
+#' By default the types of the first 20 columns are printed,
+#' \code{options(readr.num_columns)} can be used to modify this (a value of 0
+#' turns off printing).
 #'
 #' @inheritParams datasource
 #' @inheritParams tokenizer_delim
@@ -45,6 +49,7 @@ NULL
 #'   the default time zone, encoding, decimal mark, big mark, and day/month
 #'   names.
 #' @param n_max Maximum number of records to read.
+#' @param guess_max Maximum number of records to use for guessing column types.
 #' @param progress Display a progress bar? By default it will only display
 #'   in an interactive session. The display is updated every 50,000 values
 #'   and will only display if estimated reading time is 5 seconds or more.
@@ -86,26 +91,26 @@ read_delim <- function(file, delim, quote = '"',
                        col_names = TRUE, col_types = NULL,
                        locale = default_locale(),
                        na = c("", "NA"), comment = "", trim_ws = FALSE,
-                       skip = 0, n_max = -1, progress = interactive()) {
+                       skip = 0, n_max = Inf, guess_max = min(1000, n_max), progress = interactive()) {
   tokenizer <- tokenizer_delim(delim, quote = quote,
     escape_backslash = escape_backslash, escape_double = escape_double,
     na = na, comment = comment, trim_ws = trim_ws)
   read_delimited(file, tokenizer, col_names = col_names, col_types = col_types,
-    locale = locale, skip = skip, comment = comment, n_max = n_max,
-    progress = progress)
+    locale = locale, skip = skip, comment = comment, n_max = n_max, guess_max =
+      guess_max, progress = progress)
 }
 
 #' @rdname read_delim
 #' @export
 read_csv <- function(file, col_names = TRUE, col_types = NULL,
                      locale = default_locale(), na = c("", "NA"), comment = "",
-                     trim_ws = TRUE, skip = 0, n_max = -1,
+                     trim_ws = TRUE, skip = 0, n_max = Inf, guess_max = min(1000, n_max),
                      progress = interactive()) {
 
   tokenizer <- tokenizer_csv(na = na, comment = comment, trim_ws = trim_ws)
   read_delimited(file, tokenizer, col_names = col_names, col_types = col_types,
-    locale = locale, skip = skip, comment = comment, n_max = n_max,
-    progress = progress)
+    locale = locale, skip = skip, comment = comment, n_max = n_max, guess_max =
+      guess_max, progress = progress)
 }
 
 #' @rdname read_delim
@@ -113,8 +118,8 @@ read_csv <- function(file, col_names = TRUE, col_types = NULL,
 read_csv2 <- function(file, col_names = TRUE, col_types = NULL,
                       locale = default_locale(),
                       na = c("", "NA"), comment = "",
-                      trim_ws = TRUE, skip = 0, n_max = -1,
-                      progress = interactive()) {
+                      trim_ws = TRUE, skip = 0, n_max = Inf,
+                      guess_max = min(1000, n_max), progress = interactive()) {
 
   if (locale$decimal_mark == ".") {
     locale$decimal_mark <- ","
@@ -125,7 +130,7 @@ read_csv2 <- function(file, col_names = TRUE, col_types = NULL,
     trim_ws = trim_ws)
   read_delimited(file, tokenizer, col_names = col_names, col_types = col_types,
     locale = locale, skip = skip, comment = comment, n_max = n_max,
-    progress = progress)
+    guess_max = guess_max, progress = progress)
 }
 
 
@@ -134,19 +139,26 @@ read_csv2 <- function(file, col_names = TRUE, col_types = NULL,
 read_tsv <- function(file, col_names = TRUE, col_types = NULL,
                      locale = default_locale(),
                      na = c("", "NA"), comment = "",
-                     trim_ws = TRUE, skip = 0, n_max = -1,
+                     trim_ws = TRUE, skip = 0, n_max = Inf, guess_max = min(1000, n_max),
                      progress = interactive()) {
 
   tokenizer <- tokenizer_tsv(na = na, comment = comment, trim_ws = trim_ws)
   read_delimited(file, tokenizer, col_names = col_names, col_types = col_types,
     locale = locale, skip = skip, comment = comment, n_max = n_max,
-    progress = progress)
+    guess_max = guess_max, progress = progress)
 }
 
 # Helper functions for reading from delimited files ----------------------------
+read_tokens <- function(data, tokenizer, col_specs, col_names, locale_, n_max, progress) {
+  if (n_max == Inf) {
+    n_max <- -1
+  }
+  read_tokens_(data, tokenizer, col_specs, col_names, locale_, n_max, progress)
+}
+
 read_delimited <- function(file, tokenizer, col_names = TRUE, col_types = NULL,
                            locale = default_locale(), skip = 0, comment = "",
-                           n_max = -1, progress = interactive()) {
+                           n_max = Inf, guess_max = min(1000, n_max), progress = interactive()) {
   name <- source_name(file)
   # If connection needed, read once.
   file <- standardise_path(file)
@@ -159,16 +171,53 @@ read_delimited <- function(file, tokenizer, col_names = TRUE, col_types = NULL,
     data <- file
   }
 
-  col_types <- col_spec_standardise(
-    data, skip = skip, comment = comment, n_max = n_max,
-    col_names = col_names, col_types = col_types,
-    tokenizer = tokenizer, locale = locale)
+  spec <- col_spec_standardise(
+    data, skip = skip, comment = comment, n = guess_max,
+    col_names = col_names, col_types = col_types, tokenizer = tokenizer,
+    locale = locale)
+
+  if (progress) {
+     print(spec, n = getOption("readr.num_columns", 20))
+  }
 
   ds <- datasource(data, skip = skip + isTRUE(col_names), comment = comment)
 
-  out <- read_tokens(ds, tokenizer, col_types, names(col_types), locale_ = locale,
+  out <- read_tokens(ds, tokenizer, spec$cols, names(spec$cols), locale_ = locale,
     n_max = n_max, progress = progress)
 
   out <- name_problems(out)
+  attr(out, "spec") <- spec
   warn_problems(out, name)
 }
+
+generate_spec_fun <- function(x) {
+  formals(x)$n_max <- 0
+  formals(x)$guess_max <- 1000
+
+  args <- formals(x)
+
+  body(x) <-
+    call("attr",
+      as.call(c(substitute(x), stats::setNames(lapply(names(args), as.symbol), names(args)))),
+      "spec")
+
+  formals(x) <- args
+
+  x
+}
+
+#' @export
+#' @rdname read_delim
+spec_delim <- generate_spec_fun(read_delim)
+
+#' @rdname read_delim
+#' @export
+spec_csv <- generate_spec_fun(read_csv)
+
+#' @rdname read_delim
+#' @export
+spec_csv2 <- generate_spec_fun(read_csv2)
+
+#' @rdname read_delim
+#' @export
+spec_tsv <- generate_spec_fun(read_tsv)