+++ tests/testthat/test-collectors.R
@@ -28,14 +28,14 @@ test_that("only accept numbers with grouping mark", {
 # Concise collectors specification ----------------------------------------
 
 test_that("_ or - skips column", {
-  out1 <- read_csv("x,y\n1,2\n3,4", col_types = "-i")
-  out2 <- read_csv("x,y\n1,2\n3,4", col_types = "_i")
+  out1 <- read_csv("x,y\n1,2\n3,4", col_types = "-i", progress = FALSE)
+  out2 <- read_csv("x,y\n1,2\n3,4", col_types = "_i", progress = FALSE)
 
   expect_equal(names(out1), "y")
   expect_equal(names(out2), "y")
 })
 
 test_that("? guesses column type", {
-  out1 <- read_csv("x,y\n1,2\n3,4", col_types = "?i")
+  out1 <- read_csv("x,y\n1,2\n3,4", col_types = "?i", progress = FALSE)
   expect_equal(out1$x, c(1L, 3L))
 })