+++ tests/testthat/test-read-csv.R
@@ -1,7 +1,7 @@
 context("read_csv")
 
 test_that("read_csv col imputation, col_name detection and NA detection works", {
-  test_data <- read_csv("basic-df.csv", col_types = NULL, col_names = TRUE)
+  test_data <- read_csv("basic-df.csv", col_types = NULL, col_names = TRUE, progress = FALSE)
   expect_equal(unname(unlist(lapply(test_data, class))),
     c("logical", "integer", "numeric", "character"))
   expect_equal(names(test_data), c("a", "b", "c", "d"))
@@ -9,34 +9,34 @@ test_that("read_csv col imputation, col_name detection and NA detection works", 
 })
 
 test_that("read_csv's 'NA' option genuinely changes the NA values", {
-  expect_equal(read_csv("a\nz", na = "z")$a, NA_character_)
+  expect_equal(read_csv("a\nz", na = "z", progress = FALSE)$a, NA_character_)
 })
 
 test_that("read_csv's 'NA' option works with multiple NA values", {
-  expect_equal(read_csv("a\nNA\n\nmiss\n13", na = c("13", "miss"))$a,
+  expect_equal(read_csv("a\nNA\n\nmiss\n13", na = c("13", "miss"), progress = FALSE)$a,
                c("NA", "", NA, NA))
 })
 
 test_that('passing character() to read_csv\'s "NA" option reads "" correctly', {
-  expect_equal(read_csv("a\nfoo\n\n", na = character())$a, c("foo", ""))
+  expect_equal(read_csv("a\nfoo\n\n", na = character(), progress = FALSE)$a, c("foo", ""))
 })
 
 test_that("passing \"\" to read_csv's 'NA' option reads \"\" correctly", {
-  expect_equal(read_csv("a\nfoo\n\n", na = "")$a, c("foo", NA))
+  expect_equal(read_csv("a\nfoo\n\n", na = "", progress = FALSE)$a, c("foo", NA))
 })
 
 test_that("read_csv's 'skip' option allows for skipping'", {
-  test_data <- read_csv("basic-df.csv", skip = 1)
+  test_data <- read_csv("basic-df.csv", skip = 1, progress = FALSE)
   expect_equal(nrow(test_data), 9)
 })
 
 test_that("read_csv's 'skip' option allows for skipping when no header row is present'", {
-  test_data <- read_csv("basic-df.csv", skip = 1, col_names = FALSE)
+  test_data <- read_csv("basic-df.csv", skip = 1, col_names = FALSE, progress = FALSE)
   expect_equal(nrow(test_data), 10)
 })
 
 test_that("read_csv's 'n_max' allows for a maximum number of records and does not corrupt any", {
-  test_data <- read_csv("basic-df.csv", n_max = 7)
+  test_data <- read_csv("basic-df.csv", n_max = 7, progress = FALSE)
   expect_equal(nrow(test_data), 7)
   expect_equal(sum(is.na(test_data)), 0)
 })
@@ -45,7 +45,7 @@ test_that("n_max also affects column guessing", {
   df <- read_csv(n_max = 1, 'x,y,z
     1,2,3
     1,2,3,4'
-  )
+  , progress = FALSE)
   expect_equal(dim(df), c(1, 3))
 })
 
@@ -54,17 +54,17 @@ test_that("can read more than 100 columns", {
   x <- as.data.frame(matrix(rbinom(300, 2, .5), nrow = 2))
   y <- format_csv(x)
 
-  expect_equal(ncol(read_csv(y)), 150)
+  expect_equal(ncol(read_csv(y, progress = FALSE)), 150)
 })
 
 test_that("encoding affects text and headers", {
-  x <- read_csv("enc-iso-8859-1.txt", locale = locale(encoding = "ISO-8859-1"))
+  x <- read_csv("enc-iso-8859-1.txt", locale = locale(encoding = "ISO-8859-1"), progress = FALSE)
   expect_identical(names(x), "fran\u00e7ais")
   expect_identical(x[[1]], "\u00e9l\u00e8ve")
 })
 
 test_that("nuls are dropped with a warning", {
-  expect_warning(x <- read_csv("raw.csv"))
+  expect_warning(x <- read_csv("raw.csv", progress = FALSE))
   expect_equal(n_problems(x), 1)
   expect_equal(x$abc, "ab")
 })
@@ -72,37 +72,37 @@ test_that("nuls are dropped with a warning", {
 # Column warnings ---------------------------------------------------------
 
 test_that("warnings based on number of columns (not output columns)", {
-  out1 <- readr::read_csv("1,2,3\n4,5,6", "z", "__i")
-  out2 <- readr::read_csv("1,2,3\n4,5,6", FALSE, cols_only(X3 = "i"))
+  out1 <- read_csv("1,2,3\n4,5,6", "z", "__i", progress = FALSE)
+  out2 <- read_csv("1,2,3\n4,5,6", FALSE, cols_only(X3 = "i"), progress = FALSE)
 
   expect_equal(n_problems(out1), 0)
   expect_equal(n_problems(out2), 0)
 })
 
 test_that("missing last field generates warning", {
-  expect_warning(out <- read_csv("a,b\n2"))
+  expect_warning(out <- read_csv("a,b\n2", progress = FALSE))
   expect_equal(problems(out)$expected, "2 columns")
 })
 
 test_that("missing lines generates warning", {
   # first
-  expect_warning(out <- read_csv("a,b\n\n\n1,2"))
+  expect_warning(out <- read_csv("a,b\n\n\n1,2", progress = FALSE))
   expect_equal(problems(out)$expected, rep("2 columns", 2))
 
   # middle
-  expect_warning(out <- read_csv("a,b\n1,2\n\n\n2,3\n"))
+  expect_warning(out <- read_csv("a,b\n1,2\n\n\n2,3\n", progress = FALSE))
   expect_equal(problems(out)$expected, rep("2 columns", 2))
 
   # last (trailing \n is ngored)
-  expect_warning(out <- read_csv("a,b\n1,2\n\n\n"))
+  expect_warning(out <- read_csv("a,b\n1,2\n\n\n", progress = FALSE))
   expect_equal(problems(out)$expected, rep("2 columns", 2))
 })
 
 test_that("extra columns generates warnings", {
-  expect_warning(out1 <- read_csv("a,b\n1,2,3\n"))
-  expect_warning(out2 <- read_csv("a,b\n1,2,3", col_types = "ii"))
-  expect_warning(out3 <- read_csv("1,2,3\n", c("a", "b")))
-  expect_warning(out4 <- read_csv("1,2,3\n", c("a", "b"), "ii"))
+  expect_warning(out1 <- read_csv("a,b\n1,2,3\n", progress = FALSE))
+  expect_warning(out2 <- read_csv("a,b\n1,2,3", col_types = "ii", progress = FALSE))
+  expect_warning(out3 <- read_csv("1,2,3\n", c("a", "b"), progress = FALSE))
+  expect_warning(out4 <- read_csv("1,2,3\n", c("a", "b"), "ii", progress = FALSE))
 
   expect_equal(problems(out1)$expected, "2 columns")
   expect_equal(problems(out2)$expected, "2 columns")
@@ -111,39 +111,39 @@ test_that("extra columns generates warnings", {
 })
 
 test_that("too few or extra col_types generates warnings", {
-  expect_warning(out1 <- read_csv("v1,v2\n1,2", col_types = "i"))
+  expect_warning(out1 <- read_csv("v1,v2\n1,2", col_types = "i", progress = FALSE))
   expect_equal(problems(out1)$expected, "1 columns")
   expect_equal(  problems(out1)$actual, "2 columns")
 
-  expect_warning(out2 <- read_csv("v1,v2\n1,2", col_types = "iii"))
+  expect_warning(out2 <- read_csv("v1,v2\n1,2", col_types = "iii", progress = FALSE))
   expect_equal(ncol(out2), 2)
 })
 
 # read_csv2 ---------------------------------------------------------------
 
 test_that("decimal mark automatically set to ,", {
-  x <- read_csv2("x\n1,23")
+  x <- read_csv2("x\n1,23", progress = FALSE)
   expect_equal(x[[1]], 1.23)
 })
 
 # Zero rows ---------------------------------------------------------------
 
 test_that("header only df gets character columns", {
-  x <- read_csv("a,b\n")
+  x <- read_csv("a,b\n", progress = FALSE)
   expect_equal(dim(x), c(0, 2))
   expect_equal(class(x$a), "character")
   expect_equal(class(x$b), "character")
 })
 
 test_that("n_max 0 gives zero row data frame", {
-  x <- read_csv("a,b\n1,2", n_max = 0)
+  x <- read_csv("a,b\n1,2", n_max = 0, progress = FALSE)
   expect_equal(dim(x), c(0, 2))
-  expect_equal(class(x$a), "integer")
-  expect_equal(class(x$b), "integer")
+  expect_equal(class(x$a), "character")
+  expect_equal(class(x$b), "character")
 })
 
 test_that("empty file with col_names and col_types creates correct columns", {
-  x <- read_csv(datasource_string("", 0), c("a", "b"), "ii")
+  x <- read_csv(datasource_string("", 0), c("a", "b"), "ii", progress = FALSE)
   expect_equal(dim(x), c(0, 2))
   expect_equal(class(x$a), "integer")
   expect_equal(class(x$b), "integer")
@@ -153,29 +153,29 @@ test_that("empty file with col_names and col_types creates correct columns", {
 # Comments ----------------------------------------------------------------
 
 test_that("comments are ignored regardless of where they appear", {
-  out1 <- read_csv('x\n1#comment',comment = "#")
-  out2 <- read_csv('x\n1#comment\n#comment', comment = "#")
-  out3 <- read_csv('x\n"1"#comment', comment = "#")
+  out1 <- read_csv('x\n1#comment',comment = "#", progress = FALSE)
+  out2 <- read_csv('x\n1#comment\n#comment', comment = "#", progress = FALSE)
+  out3 <- read_csv('x\n"1"#comment', comment = "#", progress = FALSE)
 
   expect_equal(out1$x, 1)
   expect_equal(out2$x, 1)
   expect_equal(out3$x, 1)
 
-  expect_warning(out4 <- read_csv('x,y\n1,#comment', comment = "#"))
+  expect_warning(out4 <- read_csv('x,y\n1,#comment', comment = "#", progress = FALSE))
   expect_equal(out4$y, NA_character_)
 })
 
 test_that("escaped/quoted comments are ignored", {
   out1 <- read_delim('x\n\\#', comment = "#", delim = ",",
-    escape_backslash = TRUE, escape_double = FALSE)
-  out2 <- read_csv('x\n"#"', comment = "#")
+    escape_backslash = TRUE, escape_double = FALSE, progress = FALSE)
+  out2 <- read_csv('x\n"#"', comment = "#", progress = FALSE)
 
   expect_equal(out1$x, "#")
   expect_equal(out2$x, "#")
 })
 
 test_that("leading comments are ignored", {
-  out <- read_csv("#a\n#b\nx\n1", comment = "#")
+  out <- read_csv("#a\n#b\nx\n1", comment = "#", progress = FALSE)
 
   expect_equal(ncol(out), 1)
   expect_equal(out$x, 1L)
@@ -183,7 +183,7 @@ test_that("leading comments are ignored", {
 
 test_that("skip respects comments", {
   read_x <- function(...) {
-    read_csv("#a\nb\nc", col_names = FALSE, ...)[[1]]
+    read_csv("#a\nb\nc", col_names = FALSE, ..., progress = FALSE)[[1]]
   }
 
   expect_equal(read_x(), c("#a", "b", "c"))
@@ -193,5 +193,5 @@ test_that("skip respects comments", {
 })
 
 test_that("read_csv returns an empty data.frame on an empty file", {
-   expect_equal(read_csv("empty-file"), tibble::data_frame())
+   expect_equal(read_csv("empty-file", progress = FALSE), tibble::data_frame())
 })